pyproject.toml
[project]
name = "intern-task-bdl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.7.4",
    "fastapi>=0.115.12",
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.2",
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "langchain-core>=0.3.60",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.0.20",
    "matplotlib>=3.10.3",
    "numpy>=1.21.0",
    "ollama>=0.1.6",
    "openpyxl>=3.1.2",
    "pandas>=2.1.2",
    "pdfplumber>=0.11.6",
    "pydantic>=2.4.2",
    "pymupdf>=1.23.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
    "sentence-transformers>=2.2.2",
    "sqlalchemy>=2.0.23",
    "streamlit>=1.28.0",
    "transformers>=4.35.0",
    "uvicorn>=0.34.2",
    "watchdog>=3.0.0",
    "xlsxwriter>=3.2.3",
]
